+++ NEWS.md
@@ -10,6 +10,7 @@
 - `view_artefacts()`Shiny viewer for channel and epoch stats added.
 - `plot_timecourse()` now takes a mapping argument, which allows use of `ggplot2` `aes()` mappings
 - `eeg_average.eeg_tfr()` now follows behaviour of other `eeg_average()` methods in respecting the `epochs` structure.
+- `plot_tfr()` now always drops NA/NaN values and averages appropriately over electrodes and conditions.
 
 ### Internal changes / bug fixes
 - Now using `whitening` package for whitening before SOBI ICA
@@ -29,6 +30,7 @@
 - cleaner code in `topoplot()` for biharmonic smooth
 - `compute_psd()` now demeans individual segments when doing Welch FFT; also no longer errors when only one segment per channel
 - `eeg_tfr()` internal structure modified to keep 4 dimensions even after averaging, for consistency
+- `stat_summary_by_fill()` added to do averaging for raster plots effectively.
 
 # eegUtils 0.5.0